requirements.txt
galois==0.4.2
llvmlite==0.43.0
numba==0.60.0
numpy==2.0.2
pycryptodome==3.21.0
typing_extensions==4.12.2
